split CRDTExtraMetaLayer into multiple files
