+++ collabs/src/crdts/crdt-runtime/crdt-extra-meta-layer/index.ts
@@ -1,0 +1,2 @@
+// Only export CrdtExtraMetaLayer; rest are utilities for it.
+export * from "./crdt_extra_meta_layer";
