+++ collabs/src/crdts/crdt-runtime/index.ts
@@ -1,4 +1,4 @@
 export * from "./crdt_app";
-export * from "./crdt_extra_meta_layer";
+export * from "./crdt-extra-meta-layer";
 export * from "./crdt_extra_meta";
 export * from "./crdt_runtime";
